Renamed class AuthTicketResource to CasAuthTicketResource Changed route /auth/ticket to /v1/auth/ticket Added path ^/v1/auth/ticket? to the list of exclusions for filter CasApiAuthFilter Removed test methods from CasClient
