+++ roda-ui/roda-wui/src/main/resources/config/roda-wui.properties
@@ -42,7 +42,7 @@ ui.filter.CasWebAuthFilter.casServerUrlPrefix = https://localhost:8443/cas
 
 # ui.filter = CasApiAuthFilter
 ui.filter.CasApiAuthFilter.casServerUrlPrefix = https://localhost:8443/cas
-ui.filter.CasApiAuthFilter.exclusions = ^/swagger.json,^/v1/theme/?
+ui.filter.CasApiAuthFilter.exclusions = ^/swagger.json,^/v1/theme/?,^/v1/auth/ticket?
 
 # ui.filter = CASAuthenticationFilter
 ui.filter.CASAuthenticationFilter.casServerLoginUrl = https://localhost:8443/cas/login
